requirements.txt
pre-commit==4.2.0
pytest==8.4.0
python-dotenv==1.1.0
requests==2.32.4
Faker==37.4.0
jsonschema==4.24.0
pytest-html==4.1.1
